+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMZ48P7PXC2XETG05QQ661SW"
+        "01KN1AJRSG8FMFM6J2DYTZ81FT"
     )
     r"""The unique identifier of the data source"""