+++ packages/orq-rc/RELEASES.md
@@ -6719,3 +6719,13 @@ Based on:
 - [python v4.7.0-rc.31] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.31] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.31 - packages/orq-rc
+
+## 2026-03-31 06:54:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.32] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.32 - packages/orq-rc